Remove ceremonies-debtor-list and use ceremonies-debtor-list-import
