--- strapi/src/plugins/ceremonies-debtor-list/.nvmrc
@@ -1,1 +1,0 @@
-18.18.2
