--- strapi/src/plugins/ceremonies-debtor-list/admin/src/pages/App.tsx
@@ -1,26 +1,0 @@
-/**
- *
- * This component is the skeleton around the actual pages, and should only
- * contain code that should be seen on all pages. (e.g. navigation bar)
- *
- * Generated by Strapi plugin generator.
- */
-
-import React from "react";
-import { Switch, Route } from "react-router-dom";
-import { NotFound } from "@strapi/helper-plugin";
-import pluginId from "../pluginId";
-import HomePage from "./HomePage";
-
-const App = () => {
-  return (
-    <div>
-      <Switch>
-        <Route path={`/plugins/${pluginId}`} component={HomePage} exact />
-        <Route component={NotFound} />
-      </Switch>
-    </div>
-  );
-};
-
-export default App;
