--- strapi/src/plugins/ceremonies-debtor-list/admin/src/pluginId.ts
@@ -1,11 +1,0 @@
-/**
- * Generated by Strapi plugin generator.
- */
-import pluginPkg from "../../package.json";
-
-const pluginId = pluginPkg.name.replace(
-  /^(@[^-,.][\w,-]+\/|strapi-)plugin-/i,
-  ""
-);
-
-export default pluginId;
