--- strapi/src/plugins/ceremonies-debtor-list/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "ceremonies-debtor-list",
-  "version": "0.0.0",
-  "description": "This is the description of the plugin.",
-  "strapi": {
-    "name": "ceremonies-debtor-list",
-    "description": "Description of ceremonies-debtor-list plugin",
-    "kind": "plugin"
-  },
-  "dependencies": {
-    "moment": "^2.29.4",
-    "moment-timezone": "^0.5.37",
-    "react": "^18.2.0",
-    "xlsx": "^0.18.5"
-  },
-  "devDependencies": {
-    "@strapi/strapi": "4.11.4",
-    "@types/react": "^18.2.0",
-    "@types/web": "0.0.203"
-  },
-  "author": {
-    "name": "A Strapi developer"
-  },
-  "maintainers": [
-    {
-      "name": "A Strapi developer"
-    }
-  ],
-  "engines": {
-    "node": ">=18.x.x",
-    "npm": ">=9.8.1"
-  },
-  "scripts": {
-    "develop": "tsc -w",
-    "build": "tsc"
-  },
-  "license": "MIT"
-}
