--- strapi/src/plugins/ceremonies-debtor-list/server/config/index.ts
@@ -1,4 +1,0 @@
-export default {
-  default: {},
-  validator() {},
-};
