--- strapi/src/plugins/ceremonies-debtor-list/server/content-types/index.ts
@@ -1,2 +1,0 @@
-// Unfortunately, the content types must be created in the main Strapi codebase as I was able to set roles for their API routes if created here.
-export default {};
