--- strapi/src/plugins/ceremonies-debtor-list/server/destroy.ts
@@ -1,3 +1,0 @@
-import { Strapi } from "@strapi/strapi";
-
-export default ({ strapi }: { strapi: Strapi }) => {};
