--- strapi/src/plugins/ceremonies-debtor-list/server/index.ts
@@ -1,23 +1,0 @@
-import register from "./register";
-import bootstrap from "./bootstrap";
-import destroy from "./destroy";
-import config from "./config";
-import contentTypes from "./content-types";
-import controllers from "./controllers";
-import routes from "./routes";
-import middlewares from "./middlewares";
-import policies from "./policies";
-import services from "./services";
-
-export default {
-  register,
-  bootstrap,
-  destroy,
-  config,
-  controllers,
-  routes,
-  services,
-  contentTypes,
-  policies,
-  middlewares,
-};
