--- strapi/src/plugins/ceremonies-debtor-list/server/middlewares/index.ts
@@ -1,1 +1,0 @@
-export default {};
