--- strapi/src/plugins/ceremonies-debtor-list/server/register.ts
@@ -1,3 +1,0 @@
-import { Strapi } from "@strapi/strapi";
-
-export default ({ strapi }: { strapi: Strapi }) => {};
