--- strapi/src/plugins/ceremonies-debtor-list/strapi-server.js
@@ -1,3 +1,0 @@
-"use strict";
-
-module.exports = require("./dist/server");
