--- strapi/src/plugins/ceremonies-debtor-list/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "extends": "@strapi/typescript-utils/tsconfigs/server",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": ".",
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "jsx": "preserve",
-  },
-  "include": [
-    // Include the root directory
-    "server",
-    "admin/",
-    // Force the JSON files in the src folder to be included
-    "server/**/*.json"
-  ],
-  "exclude": [
-    "node_modules/",
-    "dist/",
-    // Do not include test files
-    "**/*.test.ts"
-  ]
-}
